Added Split the answers code into a seperate Answers code. Use Hook useRef to store non-userInterface data in a state saving container that don't refresh the entire Component. UseCallback Hook for functions.
Everything that's happened so far is as I would have but at an earlier state.
The reason the tutor did this late was to highlight the use of the "key" attribute in forcing rendering of the specific component in the list.

Note the useRef HooK: store data but doesn't trigger refresh of the User Interface.
    const shuffledAnswers = useRef();
    shuffledAnswers.current = [...allAnswers];

Note the useCallback Hook: Offers Your function will run only when the dependancy changes
    const handleSelectAnswer = useCallback(
        (selectedAnswer) => {
        setAnswerState('answered');
        setUserAnswers((prevUserAnswers) => {
            return [...prevUserAnswers, selectedAnswer];
        });

        setTimeout(() => {
            if (
            selectedAnswer === Dummy_Questions[activeQuestionIndex].answers[0]
            ) {
            setAnswerState('correct');
            } else {
            setAnswerState('wrong');
            }

            setTimeout(() => {
            setAnswerState('');
            }, 2000);
        }, 1000);
        },
        [activeQuestionIndex]
    );

    const handleSkipAnswer = useCallback(
        () => handleSelectAnswer(null),
        [handleSelectAnswer]
    );

Been a while since I worked on this. I will need to review the theory behind these 2 hooks, can't recall the rest of the reason behind them.